--- web/scenes/Portal/Teams/TeamId/Apps/AppId/Configuration/BasicInformation/graphql/client/update-app.graphql
@@ -1,11 +1,0 @@
-mutation UpdateAppInfo(
-  $app_metadata_id: String!
-  $input: app_metadata_set_input
-) {
-  update_app_metadata_by_pk(
-    pk_columns: { id: $app_metadata_id }
-    _set: $input
-  ) {
-    id
-  }
-}
